Move e2e tests for metrics export to integration status to reduce crosstalk flakiness
