+++ test/presubmit-tests.sh
@@ -28,7 +28,9 @@ source $(dirname $0)/../vendor/knative.dev/test-infra/scripts/presubmit-tests.sh
 
 # TODO(#17): Write integration tests.
 
-# We use the default build, unit and integration test runners.
+# We use the integration test runners, but need to customize build and unit
+# tests because pkg contains a number of tools and frameworks that operate at
+# the build stage that need testing.
 
 function pre_build_tests() {
   # Test the custom code generators. This makes sure we can compile the output
